requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pyautogui>=0.9.53
pygetwindow>=0.0.9
Pillow>=9.0.0
pyobjc>=8.5  # For macOS Quartz access
